ANDROID: clock_gettime(CLOCK_BOOTTIME,) slows down >20x
clock_gettime(CLOCK_BOOTTIME,) slows down after significant
accumulation of suspend time creating a large offset between it and
CLOCK_MONOTONIC time.  The __iter_div_u64_rem() is only for the usage
of adding a few second+nanosecond times and saving cycles on more
expensive remainder and division operations, but iterates one second
at a time which quickly goes out of scale in CLOCK_BOOTTIME's case
since it was specified as nanoseconds only.

The fix is to split off seconds from the boot time and cap the
nanoseconds so that __iter_div_u64_rem does not iterate.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 72406285
Change-Id: Ia647ef1e76b7ba3b0c003028d4b3b955635adabb

	modified:   arch/arm/include/asm/vdso_datapage.h
	modified:   arch/arm/kernel/vdso.c
	modified:   arch/arm64/include/asm/vdso_datapage.h
	modified:   arch/arm64/kernel/vdso.c
	new file:   arch/arm64/kernel/vdso32/compiler.h
	modified:   lib/vdso/vgettimeofday.c